[data-home] Introduce models and repository logic
This commit introduces the following changes:
* Entities and DAOs for the data that will be featured on the home screen
* Service calls to the endpoints, to fetch the data
* Repository logic to retrieve the data

+++ data/src/main/java/app/cookery/Entry.kt
@@ -1,11 +1,11 @@
 package app.cookery
 
-import app.cookery.data.entity.CookeryEntity
+import app.cookery.data.entities.CookeryEntity
 
 interface Entry : CookeryEntity {
-    val recipeId: Long
+    val recipeId: Int
 }
 
 interface MultipleEntry : Entry {
-    val otherRecipeId: Long
+    val otherRecipeId: Int
 }
